Factor out get_eventengine_output
